fix: fail better when onboard fails (#18)
* fix: fail better when onboard fails

Making sure we properly error out when something happens. The way it was
before basically masked any potential error and for something like
systemd, it made it impossible to restart on failure, as it is what we
want and onboarding should retry until to2 done (maybe to0 hasn't been
started yet but the device got powered on etc etc).

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

* fix(onboard): get rid of rvonly

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

---------

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>